+++ pyogrio/geopandas.py
@@ -16,6 +16,7 @@ def read_dataframe(
     where=None,
     bbox=None,
     fids=None,
+    fid_as_index=False,
 ):
     """Read from an OGR data source to a GeoPandas GeoDataFrame or Pandas DataFrame.
     If the data source does not have a geometry column or ``read_geometry`` is False,
@@ -26,7 +27,7 @@ def read_dataframe(
     Parameters
     ----------
     path : str
-        path to file
+        A dataset path or URI.
     layer : int or str, optional (default: first layer)
         If an integer is provided, it corresponds to the index of the layer
         with the data source.  If a string is provided, it must match the name
@@ -67,6 +68,9 @@ def read_dataframe(
         specific (e.g. typically 0 for Shapefile and 1 for GeoPackage, but can
         still depend on the specific file). The performance of reading a large
         number of features usings FIDs is also driver specific.
+    fid_as_index : bool, optional (default: False)
+        If True, will use the FIDs of the features that were read as the
+        index of the GeoDataFrame.  May start at 0 or 1 depending on the driver.
 
     Returns
     -------
@@ -81,13 +85,7 @@ def read_dataframe(
     except ImportError:
         raise ImportError("geopandas is required to use pyogrio.read_dataframe()")
 
-    path = str(path)
-
-    if not "://" in path:
-        if not "/vsi" in path.lower() and not os.path.exists(path):
-            raise ValueError(f"'{path}' does not exist")
-
-    meta, geometry, field_data = read(
+    meta, index, geometry, field_data = read(
         path,
         layer=layer,
         encoding=encoding,
@@ -99,11 +97,17 @@ def read_dataframe(
         where=where,
         bbox=bbox,
         fids=fids,
+        return_fids=fid_as_index,
     )
 
     columns = meta["fields"].tolist()
     data = {columns[i]: field_data[i] for i in range(len(columns))}
-    df = pd.DataFrame(data, columns=columns)
+    if fid_as_index:
+        index = pd.Index(index, name="fid")
+    else:
+        index = None
+
+    df = pd.DataFrame(data, columns=columns, index=index)
 
     if geometry is None or not read_geometry:
         return df